Only reset settings when a gemfile is already set
Only thing we need to reset is settings, and only because settings are
loaded relative to the `Gemfile`, so they need to be reset if the
configured or passed `Gemfile` is not standard and will resolve to a
different root.

Resetting more stuff causes issues under some circumstances,
specifically when dependencies further tweak `Kernel.require`.